+++ entities/nodes/Node.h
@@ -34,7 +34,7 @@ public:
   virtual ~Node();
 
   // If NodeAnimation is a nullptr, bone matrix data is used instead of animation data (i.e. default bone pose is used)
-  virtual void update_armature(glm::mat4 parent_transformation, RootNode* root_node, NodeAnimation* animation, int animation_time);
+  virtual void update_armature(glm::mat4 parent_transformation, RootNode* root_node, NodeAnimation* animation, float animation_time);
   virtual void draw(Shader_Opacity_Triplet shaders, Shader::DrawType draw_type, std::vector<Transparent_Draw>* partially_transparent_meshes=nullptr, glm::mat4 model=glm::mat4(1.0f), int texture_unit=0);
 
   // Getters & setters
